+++ oauth2_provider/models.py
@@ -12,7 +12,7 @@ from django.utils.translation import ugettext_lazy as _
 from .generators import generate_client_id, generate_client_secret
 from .scopes import get_scopes_backend
 from .settings import oauth2_settings
-from .validators import validate_uris
+from .validators import RedirectURIValidator, WildcardSet
 
 
 class AbstractApplication(models.Model):
@@ -65,7 +65,6 @@ class AbstractApplication(models.Model):
 
     redirect_uris = models.TextField(
         blank=True, help_text=_("Allowed URIs list, space separated"),
-        validators=[validate_uris]
     )
     client_type = models.CharField(max_length=32, choices=CLIENT_TYPES)
     authorization_grant_type = models.CharField(
@@ -125,12 +124,29 @@ class AbstractApplication(models.Model):
 
     def clean(self):
         from django.core.exceptions import ValidationError
-        if not self.redirect_uris \
-            and self.authorization_grant_type \
-            in (AbstractApplication.GRANT_AUTHORIZATION_CODE,
-                AbstractApplication.GRANT_IMPLICIT):
-            error = _("Redirect_uris could not be empty with {grant_type} grant_type")
-            raise ValidationError(error.format(grant_type=self.authorization_grant_type))
+
+        grant_types = (
+            AbstractApplication.GRANT_AUTHORIZATION_CODE,
+            AbstractApplication.GRANT_IMPLICIT,
+        )
+
+        redirect_uris = self.redirect_uris.strip().split()
+        allowed_schemes = set(s.lower() for s in self.get_allowed_schemes())
+
+        if redirect_uris:
+            validator = RedirectURIValidator(WildcardSet())
+            for uri in redirect_uris:
+                validator(uri)
+                scheme = urlparse(uri).scheme
+                if scheme not in allowed_schemes:
+                    raise ValidationError(_(
+                        "Unauthorized redirect scheme: {scheme}"
+                    ).format(scheme=scheme))
+
+        elif self.authorization_grant_type in grant_types:
+            raise ValidationError(_(
+                "redirect_uris cannot be empty with grant_type {grant_type}"
+            ).format(grant_type=self.authorization_grant_type))
 
     def get_absolute_url(self):
         return reverse("oauth2_provider:detail", args=[str(self.id)])